requirements.txt
gensim==4.3.3
joblib==1.4.2
kafka-python==2.0.2
networkx==3.4.2
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.5.1
scipy==1.13.1
six==1.17.0
smart-open==7.1.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2025.1
wrapt==1.17.2
